+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 - [2021-01-05] Add Angular and Vue sites
   [\#142](https://github.com/thisdot/framework.dev/pull/142)
+  ([jbachhardie](https://github.com/jbachhardie))
+- [2022-01-05] Resolve known bugs created with merge of homepage content
+  [\#150](https://github.com/thisdot/framework.dev/pull/150)
+  ([tvanantwerp](https://github.com/tvanantwerp))
 - [2022-01-04] Fix trailing slashes on URLs
   [\#147](https://github.com/thisdot/framework.dev/pull/147)
   ([jbachhardie](https://github.com/jbachhardie))